Fail the Coverity build if their script cannot be downloaded
If curl encounters an error trying to download the Coverity build script
(as seems to be happening at the moment due to SSL issues), we will not
be notified since the pipe swallows the error code. This is not ideal, so
this commit updates the Travis build steps to use a slightly more contrived
construction which will get and then execute the script as independent steps.
Also, to ensure curl returns an error code even in less-severe cases (e.g. a
basic 404 "Not Found") we add the '-f' option.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>