Use wcmScrollDistance as scroll threshold; recognize scrolling more consistently
The wcmFingerScroll function runs a series of checks when trying to
recognize a scroll gesture. The first check is that the fingers have
not spread apart (or come closer together) by more than a specified
distance (wcmMaxScrollFingerSpread). The next check is that both
fingers are moving in the same direction. This second check is performed
by 'pointsInLine', which returns 'TRUE' if the contact has moved in the
expected direction by some minimum distance. This minimum distance was also
set to wcmMaxScrollFingerSpread, despite it not really being a measure of
finger spreading.

This commit changes the minimum distance to be wcmScrollDistance: the
distance at which a scroll event would normally be emitted. This will
cause scrolling recognition to feel consisitent with the rest of the
scrolling experience, eliminating the lag between starting a scroll
gesture and the driver recognizing it. As long as you move your fingers
one scroll distance, scrolling will begin.

Users who have a very small scroll distance may find this change makes
it harder to trigger a pinch zoom since the fingers must move apart or
together by at least wcmMaxScrollFingerSpread before moving up/down/left/right
by the scroll distance.

Ref: https://github.com/linuxwacom/xf86-input-wacom/issues/43
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>