Do not wait wcmTapTime to enter scroll or zoom mode
Commit 3005fc0 reduced the amount of time that we wait before entering
scroll and zoom mode, mentioning that the only reason we want to wait
at all is to process time-based two-finger gestures. The commit hints
at a "2 finger tap" being the only gesture which this would apply to,
but there are no signs of that gesture in the driver anymore. There is
the right-click gesture which requires that the second finger be tapped
for less than wcmTapTime, but removing the wcmTapTime limit on entering
scroll and zoom modes does not affect that gesture (unless your right-
click gesture was so sloppy that your fingers moved enough to be
considered a scroll or zoom gesture).

Ref: https://github.com/linuxwacom/xf86-input-wacom/issues/43
Ref: 3005fc0 ("reduce 2-finger scroll holdoff time")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>