Remember the event types we receive and skip events with no data
On RHEL7.x kernels we get event frames with merely MSC_SERIAL -1 for some
devices on proximity in. This is caused by the accelerometer data that is
otherwise suppressed by those kernels.

E: 123.456 0000 0000 0000       # ------------ SYN_REPORT (0) ----------
E: 123.456 0004 0000 -001       # EV_MSC / MSC_SERIAL           -1

For a MSC_SERIAL -1 we default to the PAD_ID (0x10), despite the events
happening on the Pen device node. This triggers an error message during EV_SYN
processing:

   (EE) usbDispatchEvents: Device Type mismatch - 16 -> 0. This is a BUG.

Once we receive the BTN_TOOL_PEN when the actual pen comes into proximity, the
error message goes away because our tool type matches with what we expect.

Fix this issue by remembering which event types we received in the current
frame. If all we got was EV_MSC, skip the event dispatch - we don't have any
data to process anyway.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>