Add feature flags for several missing tablets
Feature flags are used by the driver to determine if a particular device
needs special treatment. The devices usually work "okay" without these
flags set, but missing flags can result in some features not working
or not working as expected. For instance, the touch ring on the Bluetooth
interface of the 2nd-gen Intuos Pro does not work at the moment because
the WCM_RING feature is missing. It also gains the WCM_ROTATION flag.
All other devices added gain the WCM_LCD flag, as well as the WCM_ROTATION
flag if the device supports pen rotation.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Aaron Armstrong Skomra <aaron.skomra@wacom.com>