Fix panscroll when using non-default tablet area
Attempting to perform a panscroll operation on a tablet that has a non-
default area set can cause scrolling to occur even when the pen isn't
moved. The direction of scrolling may change depending on the position
of the active area (e.g. scrolling down when the area is on the top;
scrolling up when the area is on the bottom).

The cause is a result of the current and previous DeviceState structures
(`ds` and `priv->oldState`) containing different information about the
pen location. While the current state contains the untransformed device
coordinates, when `wcmUpdateOldState()` is called, the XY information is
replaced with the transformed version.

To fix this issue, we modify the wcmPanscroll function to take X and Y
as explicit parameters that are obtained from the valuators. These are
transformed values and so will be compatible with changes to the area.
Because the valuator data may have been transformed to a delta within
wcmSendNonPadEvents we also need to add a check if the device is in
relative mode and then act appropriately.

Ref: https://github.com/linuxwacom/xf86-input-wacom/issues/59
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>