Use a proxout height of 30 for all stylus devices
Styli and pucks likely have different proxout height requirements due
to differences in how the tools are held/used. For now, lets try setting
a height of 30. This should be close to the height previously present for
consumer devices (before the kernel started honoring prox/range) and
should hopefully be reasonable for professional devices too.

Ref: https://github.com/linuxwacom/input-wacom/issues/54
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>