Handle multitouch mode up at the very start wcmGestureFilter
Moving the code which handles multitouch mode to the start of the function
both makes the code more obvious and ensures that we don't accidentally
recognize an unwanted gesture. Doing so much partial-processing before
eventually hitting a "goto ret" and handling multitouch mode was tricky
to understand and hard to verify.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>