Increase size of event queue to 128
The size of the event queue defines how many kernel events we can keep
track of before seeing a SYN_REPORT. If we are ever asked to handle
more than this limit, data is dropped with unpredictable consequences.
Our previous queue size of 32 was usually sufficient but could still
be bumped into in edge-cases. Updating 8 or more touch contacts at the
same time could sometimes cause the queue to run out of space, for
example.

This commit bumps the queue size up to 128. This should be more than
enough for all current devices (some of which can track up to 16 contacts).

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>