Add workflow for automated testing via Github Actions
We've been using Travis for our automated tests, but that service is
becoming more difficult to use (open source projects have to periodically
request free credits to keep running). Github itself can be used as a
replacement by using its "Actions" infrastructure to execute various
defined workflows.

This commit ports the Travis script into Github syntax. Most everything
was translated without issue, but support for the ppc64le architecture
is notably missing. This is not a critical target for us, however, so
we ignore it for the time being.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>