A channel with serial 1 is our current channel
Where MSC_SERIAL comes in late, we already created a channel for this
tool with the default serial 1, see protocol5Serial(). Thus if we now
have an event for the same device type, this time with an actual serial
number, let's re-use that channel.

Otherwise, a MSC_SERIAL event causes a new channel to be created and we
lose our DeviceState information (it being memset).

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>