Simplify axis resolution initialization a bit
No-one really cares about min/max resolution anyway, so let's simplify
this: if the resolution is zero, set it all to 1 (server requirement),
otherwise set min 0 and max whatever the resolution is. This almost
matches what we had except for tilt where we had min_res == max_res. but
as I said, no-one cared about that.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>