Abstract initializing the device into several helpers
Define an enum for the various axis types we have and initialize those
with those names. This is largely just moving code around but it does
make things a bit easier to read, in particular we separate setting the
axis ranges (generic) from the assigned property name and valuator
number (X specific).

The biggest difference: we now hardcode Absolute as mode on each axis on
the device. Functionally identical to before though.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>