Abstract the event interface to pass a struct with axis data around
Instead of the first_valuator/num_valuators/vararg combo, put the
various axis values into a struct with defined names for each field and
a mask that tells us which values are currently set.

Pass that struct around and just before we get to the X server, convert
that struct into the first_valuator/num_valuators/vararg combo. This
is basically what we already did for the device state anyway, it just
moves the conversion a few layers higher.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>