Use WacomDevicePtr as primary context argument
This changes all functions (wherever possible) to take a WacomDevicePtr
instead of an InputInfoPtr as context argument, with the InputInfoPtr
either assigned to priv->pInfo or removed where now superfluous.

This should have no functional changes.

Remaining functions are those required by the X server to have a certain
signature.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>