Add the WacomInterface header file
This header file now declares the functions required by any
frontend implementation of this driver. The core part of the driver (all
files but xf86Wacom.c) will eventually be independent of X altogether,
with the X driver bit in xf86Wacom.c the layer that connects the wacom
driver to the X server - with the APIs the server requires.

This interface is currently semi-complete, we still rely on pInfo in the
driver and we still use several X-specific interfaces but the remaining
ones are ones that require simple 1:1 conversion but not functional
changes (e.g. option handling, fd retrieval, etc.)

Since Bool is not used anywhere but in X, switch to bool instead.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>