Declare the three special property update functions
These are updated from the core driver and call back into the property
code. ATM there are no plans to make the property code X independent, so
it's easiest to just declare the functions here and leave them to the
frontend to implement (as stubs, usually).

Because of SIGIO restrictions in older X servers, these need to be
handled through a timer so we leave the implementation in wcmXCommand.c
- link that file into the driver and it all comes for free.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>