Add a GObject-based interface to the driver
This patch adds a basic GObject-based library in lib/ that is sufficient
to interact with the driver. The example tools/wacom-record.c is
provided to illustrate the use of this library and to provide a
commandline debugging tool.

This library is for testing use only, it does not provide a stable API
and should not be used. It is not installed and only available if the
driver is built with meson.

GObject IR bindings are available too to write e.g. Python programs that
use the driver.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>