Remove support for serial devices from the driver
The serial code inside the driver is effectively unused and definitely
untested. It's been a decade since the last serial devices came out and few of
those will be in active use, the few that are are likely supported by
the inputattach code we keep in place for now.

Fixes #197

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>