meson: default to the module dir from xorg-server.pc
On Ubuntu 20.04 at least, libdir is /usr/lib/x86_64-linux-gnu/ but the
xorg modules are in /usr/lib/xorg/modules/. A default install with
prefix /usr thus puts the modules into the wrong directory.

Fix this by defaulting to the xorg module path from the xserver
pkgconfig file. That has been exported since 2005, so it's save to rely
on.

Fixes #217

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>