Use locally-declared unsigned loop variables to eliminate warnings
We often use the prototypical `int i` loop variable but sometimes we're
iterating over a loop with an unsigned bound. This is low-hanging fruit
for cleaning up some of the "Wsign-compare" warnings that exist in the
driver. While we're at it, take the opportunity to make many of these
loops use local declarations -- a feature standardized way back in
1999 alongside other blockbusters like The Matrix.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>