Replace our unit-test suite with something more magic and easier
The previous approach, written 11 years ago, relied on providing all
symbols the driver needs as stubs (see fake-symbols.c). Most of these
symbols just needed to be there so we could compile the test, they
didn't actually get utilized by the tests.

Dropping this and instead add something roughly similar to Rust: the
test function can be in the same file as the source function.

The TEST_CASE() macro uses __attribute__((section)) to push the test
cases in custom ELF section. This section we can iterate on and call all
tests one-by-one. The test runner only needs to dlopen + dlsym the
driver module and run the entry point function.

The test-suite part of this is minimal: tests can assert on failure and
everything stops. Adding things like fnmatch for test case names is
possible when/if we ever need it.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>