Rotate / scale coordinates before storing them in axes
The WacomAxisData calculated by `wcmSendEvents` is expected to contain
the X,Y coordinates which have been adjusted for device rotation and
scale. Prior to the refactoring, the values of X and Y (along with
everything else) were only packaged up for relaying at the very end
of the function. Now this packaging occurs much earlier which requires
our call to `wcmRotateAndScaleCoordinates` to move as well.

Fixes: 58a931bc21d1 (Abstract the event interface to pass a struct with axis data around )
Link: https://github.com/linuxwacom/xf86-input-wacom/issues/226
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>