Revamp debug printing of axis data
The `wcmSendEvents` function is a bit of a mess when it comes to printing
out debug information. It maintains and dumps temporary data that may not
reflect what will actually be passed to e.g. `wcmSendNonPadEvents` and
does so at two different log levels for little reason after the recent
refactor.

This commit takes a page from wacom-record and introduces a function to
dynamically dump axis data to a string. This lets us remove all the old
temporaries and print out the values that actually matter.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>