gwacom: make the enums GIR-compatible
GIR generation requires "typedef enum {} foo", with the first character
of the enum names dropped. In the bindings, WFOO_BAR then turns into
FOO_BAR on the enum object.

This fixes an ABI issue with the WacomAxis struct - the enum (4 bytes)
was previously exported as gpointer (8 bytes) in the bindings, causing
garbage values to be read.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>