test: mark the test section entries as no_sanitize_address
This works around an issue running our test suite through clang with
asan. Specifically, the asan test fails with:
  ERROR: AddressSanitizer: global-buffer-overflow on address
  #0 0x7f31314a1b57 in wcm_run_tests [...] wacom-test-suite.c:44:3
  ...
  0x7f31314ff150 is located 48 bytes to the left of global variable '_decl_test_get_wheel_button' defined in '../src/wcmCommon.c:1802:1' (0x7f31314ff180) of size 16
  0x7f31314ff150 is located 0 bytes to the right of global variable '_decl_test_get_scroll_delta' defined in '../src/wcmCommon.c:1754:1' (0x7f31314ff140) of size 16

The C standard doesn't guarantee that sections are immediately
next to each other, so our expected behavior could break in the future.

Since this is a test suite, we don't really care for now. And I rather
suspect there's plenty of code out there that too would break if clang
ever changed this.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>