wacom-record: fix printing for multiple devices
wacom-record takes multiple event nodes but the device info printing was
broken. Fix this by moving the device info printing into the
device-added callback (which is emitted as soon as we call
wacom_device_new). The new event output is thus:

    - source: 2
      event: new-device
      name: "Wacom Intuos Pro M Pen"
      options:
      - _source: "server/udev"
      - device: "/dev/input/event4"
      ...

This now has the advantage of subdevices printing the correct options as
well (i.e. type eraser as this is added by the hotplugging code).

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>