test: skip the wacom tests if we can't open the .so file
Distribution package builds that use -Wl,-z,relro -Wl,-z,now will
attempt to resolve all symbols on dlopen(). This is explicitly what we
do not want because we don't link in the whole X server.

It's a bit hard to override this for a single executable (my attempts
didn't go far so there may be some other flag that influences this too)
so for now just skip the whole test - this is better than failing the
test suite.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>