scan-build: Silence dead-store warning from intentionally-ignored value
We don't care about the result from `read`, but it seems that Coverity
was complaining that we didn't use its result. We hacked around this
by simply storing it into `rc`, but now scan-build complains that this
value is also never used. This commit further-refines the hack to
silence both Coverity and scan-build.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>