test: add a test for the artpen rotation
The artpen has a physical rotation property, sent by the kernel driver
as ABS_Z (historical reasons). In the driver this is mapped to the
ds->abswheel state because this axis is shared with the airbrush wheel -
both are sent through the same XI valuator (historical reasons, as
usual).

We can rather easily test this by ensuring we have an artpen device id
and then send something through ABS_Z, that should update the wheel
axis.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Aaron Skomra <aaron.skomra@wacom.com>