Fix xsetwacom returning 0 even when there's an error
xsetwacom was only able to return a failure on a few scenarios but would return a success even when the command would encounter failures.

List, set, get, and some other functions within those all had the potential to fail and would print an error message when said failures occured but would improperly return a success.
This commit adds the ability to properly report a success(0) or a failure(-1) to the relevant functions including all get and set methods that can be called by get_func and set_func.

Link: https://github.com/linuxwacom/xf86-input-wacom/issues/126
Signed-off-by: Joshua Dickens <joshua.dickens@wacom.com>