test: Ensure standard pen does not report the "wheel" axis
The event data has a mask field that specifies which fields are actually
present (zero may be a valid value for an axis after all). Let's check
that mask for each event and ensure that if the pen supports the axis,
the axis is also present in the event.

We still send kernel events for all axes to ensure the unsupported ones
are correctly ignored by the driver.

This allows us to parametrize the test with different stylus types and
re-use the rest. Here we add a standard Cintiq pen and make sure the
wheel axis is never set.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>