CI: add a trigger job to build the test suite on gitlab
This triggers a workflow run in a custom repository that has a KVM setup
to run the tests that require uinput devices. That repo is nothing more
than a gitlab-ci.yml file that pulls our repo here and runs meson test
inside a QEMU VM. This way we can run those tests that require uinput as
well.

See the repo itself here:
https://gitlab.freedesktop.org/whot/xf86-input-wacom-ci/

Unlike the other workflows, this one only runs when pushing, not on a
PR, for two reasons:
- safer this way since we control the repo (and the secrets required are
  only set in the our repo, not a user fork)
- easier this way, running on a PR would require figuring out which
  branch of which user repo to clone and test. Probably all in the
  gitlab CI variables, but I haven't done this bit yet.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>