Bring two finger right-click function back
Patch 65b7c37 fixed spurious right-clicks by setting wcmGestureMode
back to GESTURE_CANCEL_MODE immediately when there is less than two
fingers on the tablet, if the wcmGestureMode was in SCROLL_MODE or
ZOOM_MODE.

However, it narrowed wcmFingerTapToClick() rotine to LAG_MODE, which
removed the opportunity for the routine to executed. Revert that.

Fixes: https://github.com/linuxwacom/xf86-input-wacom/commit/65b7c37f0fac5cdd03ff7f977124148d0bdf4be8
(Prevent spurious right-clicks at the end of very short scroll and zoom gestures)

Signed-off-by: Ping Cheng <ping.cheng@wacom.com>