gwacom: add support for special runtime options
This is a hack to make it possible to test options that cannot be set
through the xorg.conf, e.g button actions that are set through the
X-specific properties interface.

Expose a new function for "runtime options", though the actual option
implementation needs to be handled in the gwacom wrapper code - i.e.
only options explicitly supported can work.

The currently supported option is "PanButton" which maps the given
button to the PanScroll action.