Add option SmoothPanscrollingEnabled
This is a backwards-compatibility option similar to the Pressure2K
option. GDK2 applications are effectively limited to 7 axes
(value of GDK_AXES_LAST in GDK2) and adding smooth panscrolling gives
our device 8 axes total. This can cause issues when the driver, GDK and
the application don't agree on the number of axes in the device.

This is an application/GDK bug but as for the Pressure2K option we
expect there to be applications that cannot be updated easily. To work
around this, provide a driver option to disable this new feature
altogether and effectively return to 6 axes for the pen device.

This is a partial revert of fe923e927a "Implement smooth panscrolling"
for the implementation. Minor changes though, the implementation now
relies more on local variables than pointers.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>