release.sh: update to take a --token value
Apparently the old way no longer works, so let's update with a
copy/paste from the examples.

Intentional switch to use --token because that's how you should be doing
releases anyway - with a temporary token that can be revoked after the
release is complete (and that also encodes the username).

https://docs.github.com/en/rest/releases/releases?apiVersion=2022-11-28#create-a-release