Avoid out-of-bounds array read in usbInitToolType and fix its fallback
The usbInitToolType function uses `deriveDeviceTypeFromButtonEvent` as
a fallback when it cannot otherwise discover the type of tool being
reported in a packet of kernel events. This fallback was intended
to be called for each event in the packet by iterating over the
`event_ptr` array. The loop is incorrectly implemented, however. Before
we get to this point, the value of `event_ptr` would have already been
incremented to the end of the list by the deviceTypeFromEvent loop.
This means our calls to deriveDeviceTypeFromButtonEvent are accessing
data beyond the valid list of events and possibly even beyond the end
of the array.

To fix this, we can replace the pointer arithmetic in both loops with
array index notation instead.

Fixes: 3eacea1058ef ("Attempt to derive the tool type from a known button/key event")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>