Recognize the pad in more situations
Allow dial and ring events to be recognized as coming from the pad
when there are no other better sources of information (for example,
BTN_TOOL_* events earlier in the event packet, an ABS_MISC with the
information, or a previously-known tool type).

This fixes an issue where the pad is not recognized (and events are
not generated) if your first interaction is with a relative wheel
rather than a button or touchring. The events types matched by this
function have an ambiguous source on their own, but other earlier
tests would be able to determine the type from other context. If
all other context-sensitive tests fail to find a tool type, it is
reasonably safe to assume we're looking at a pad.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>